Revert "円を8分割する線を表示する機能を追加する"
