+++ Sources/SmallRing.swift
@@ -9,8 +9,6 @@ import SwiftUI
 
 /// Small ring
 struct SmallRing: View {
-    @EnvironmentObject var viewModel: OMJoystickViewModel
-
     var smallRingDiameter: CGFloat
     var subRingColor: Color
